Use targetgroup for alb widgets
The datadog `name` tag on albs is not unique per region. The queries
adjusted here were filtering on only the `name` tag, which would
aggregate the values from all regions where albs were deployed with this
same name. That meant we were getting global graphs for timeboards that
were supposed to be scoped to a region.

`targetgroup` tag, which we are passing in as configuration already,
*is* unique per region. Adjusted the hits, latency, and total request
widgets to use `targetgroup`.

Change-Id: Ie66ed53ac174c2dc7eb7775bb6ced0af7cfdf055
Reviewed-on: https://gerrit.instructure.com/186290
Reviewed-by: Chris Wang <cwang@instructure.com>