Add requests and jobs charts
The metrics for these charts are sourced from statsd gems. The gems can
be configured to write to a custom metric namespace. You can set the
same namespace in your dashboards with the `metric_prefix` option.

For request widgets: It's useful to exclude health checks, especially
for regions which have less traffic. You can pass in a list of Datadog
tags which will be exluded with the `excluded_tags` option.

For request widgets: The default is to add a `by {controller,action}`
clause to the query for Rails apps. If you're application is not Rails
or tagged differently you can override the by clause tag lits with the
`by_tags` option.

Change-Id: I064c0db73ec38ea8b93408703c1dde28a72cdaab
Reviewed-on: https://gerrit.instructure.com/187519
Reviewed-by: Walter Wilfinger <wwilfinger@instructure.com>
Reviewed-by: Bryan Petty <bpetty@instructure.com>